Refactor SecureConversationsCoordinator environment initialization
