+++ GliaWidgets/Sources/Coordinators/EngagementCoordinator/EngagementCoordinator.swift
@@ -437,54 +437,16 @@ extension EngagementCoordinator {
             messagingInitialScreen: messagingInitialScreen,
             viewFactory: viewFactory,
             navigationPresenter: navigationPresenter,
-            environment: .init(
+            environment: .create(
+                with: environment,
                 queueIds: interactor.queueIds,
-                listQueues: environment.listQueues,
-                sendSecureMessagePayload: environment.sendSecureMessagePayload,
-                createFileUploader: environment.createFileUploader,
-                uploadSecureFile: environment.uploadSecureFile,
-                fileManager: environment.fileManager,
-                data: environment.data,
-                date: environment.date,
-                gcd: environment.gcd,
-                createThumbnailGenerator: environment.createThumbnailGenerator,
-                uuid: environment.uuid,
-                uiApplication: environment.uiApplication,
-                uiScreen: environment.uiScreen,
-                notificationCenter: environment.notificationCenter,
-                createFileUploadListModel: environment.createFileUploadListModel,
                 viewFactory: viewFactory,
-                fetchFile: environment.fetchFile,
-                createFileDownload: environment.createFileDownload,
-                loadChatMessagesFromHistory: environment.loadChatMessagesFromHistory,
-                fetchChatHistory: environment.fetchChatHistory,
-                fetchSiteConfigurations: environment.fetchSiteConfigurations,
                 chatCall: chatCall,
                 unreadMessages: unreadMessages,
-                showsCallBubble: false,
+                showCallBubble: false,
                 screenShareHandler: screenShareHandler,
                 isWindowVisible: isWindowVisible,
-                uploadFileToEngagement: environment.uploadFileToEngagement,
-                getCurrentEngagement: environment.getCurrentEngagement,
-                submitSurveyAnswer: environment.submitSurveyAnswer,
-                interactor: interactor,
-                getSecureUnreadMessageCount: environment.getSecureUnreadMessageCount,
-                messagesWithUnreadCountLoaderScheduler: environment.messagesWithUnreadCountLoaderScheduler,
-                secureMarkMessagesAsRead: environment.secureMarkMessagesAsRead,
-                downloadSecureFile: environment.downloadSecureFile,
-                isAuthenticated: environment.isAuthenticated,
-                startSocketObservation: environment.startSocketObservation,
-                stopSocketObservation: environment.stopSocketObservation,
-                createSendMessagePayload: environment.createSendMessagePayload,
-                orientationManager: environment.orientationManager,
-                proximityManager: environment.proximityManager,
-                log: environment.log,
-                timerProviding: environment.timerProviding,
-                snackBar: environment.snackBar,
-                maximumUploads: environment.maximumUploads,
-                cameraDeviceManager: environment.cameraDeviceManager,
-                flipCameraButtonStyle: environment.flipCameraButtonStyle,
-                alertManager: environment.alertManager
+                interactor: interactor
             )
         )
